+++ src/Exception/JAQBException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace JAQB\Exception;
+
+/**
+ * The base exception class for JAQB.
+ */
+class JAQBException extends \Exception
+{
+}